requirements.txt
# ai requirements
chromadb
openai
langchain
tiktoken
unstructured
pypdf

# atlassian
atlassian-python-api
